removed dependency on Amoeba
